This is a large commit fixing the problem with DOM caching (fixes #68, #67)
Since Firefox 15, whenever a page unloads, DOM nodes created within this page
are destroyed. FloatNotes used to cache the DOM node for a specific note and
reuse it. Since this is not possible anymore, whenever notes have to be
refetched, new DOM nodes are created.
This also implies that notes themselves are not cached anymore, each load hits
the DB.

Other changes:

- Separation of URL parsers and builders
- File URLs can now be selected up to the root (fixes #57, #51)